refactor: allocate responisbility for creating instance to scanner
