+++ mvc/src/test/java/nextstep/mvc/controller/tobe/mappings/ControllerScannerTest.java
@@ -2,7 +2,7 @@ package nextstep.mvc.controller.tobe.mappings;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Set;
+import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import samples.TestController;
@@ -10,13 +10,13 @@ import samples.TestController;
 class ControllerScannerTest {
 
     @Test
-    @DisplayName("컨트롤러 어노테이션이 있는 클래스를 찾는다")
+    @DisplayName("컨트롤러 어노테이션이 있는 클래스로 인스턴스를 만든다")
     void findControllers() {
         // given
         ControllerScanner scanner = new ControllerScanner();
         // when
-        Set<Class<?>> classes = scanner.findClasses("samples");
+        Map<Class<?>, Object> instanceByClasses = scanner.createInstanceByClasses("samples");
         // then
-        assertThat(classes).containsOnly(TestController.class);
+        assertThat(instanceByClasses.keySet()).containsOnly(TestController.class);
     }
 }
